chore: replace constant file by moving items in ciConfigConstant
