+++ src/components/ciConfig/CIDockerFileConfig.tsx
@@ -16,8 +16,7 @@ import { CI_BUILDTYPE_ALIAS } from './CIConfig.utils'
 import { getBuildpackMetadata, getDockerfileTemplate } from './service'
 import { DockerConfigOverrideKeys } from '../ciPipeline/types'
 import { BuildersAndFrameworksType, CIDockerFileConfigProps, LoadingState } from './types'
-import { RootBuildContext } from './ciConfigConstant'
-import { CI_BUILD_TYPE_OPTIONS } from './constants'
+import { RootBuildContext, CI_BUILD_TYPE_OPTIONS } from './ciConfigConstant'
 import { FEATURE_DISABLED } from '../../config/constantMessaging'
 import { ReactComponent as CheckIcon } from '../../assets/icons/ic-check.svg'
 import { ReactComponent as ErrorIcon } from '../../assets/icons/ic-error-exclamation.svg'
